api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.19
pgvector>=0.2.5
redis>=5.0.6
rq>=1.16.2
python-multipart>=0.0.9
emails>=0.6
python-dotenv>=1.0.1
